Initial commit of groktools website after cloning
Why
---
Cloned existing jekyll-bootstrap website shell so
creating the groktools website becomes easy.
I've changed some text here and there to get a feel
for how the website would look like, but a lot of
text needs to be changed still. This commit is so that
i can bring all the css down locally. working on a plane
=> cannot rely on CDN :).

A couple of notes on some implementation choices: instead
of pictures, i'm going with glyphs from the fontawesome
library for my projects because i dont want to spend time
creating crappy images. Also, I've added a link for research
and taken out the contact form, etc.