Convert the boilerplate to actual groktools website.
Why
---
Since I didnt expect this jekyll template to work out actually, I kept
playing with it to see if it would. Now it seems viable enough as a
first version website. There are a lot vestiges from the template, and
some silly style attributes littered thoroughout that make this code really
not the best one in the long run, but it makes for a good starter website

Not itemising the changes and techniques therefore.